refactor: 리뷰 조회 훅 로직 공통화 및 서비스 함수 분리
