+++ src/hooks/review/useRecentReviews.js
@@ -1,28 +1,9 @@
 import { useQuery } from "@tanstack/react-query";
-import {
-  getFirestore,
-  collection,
-  query,
-  orderBy,
-  limit,
-  getDocs,
-} from "firebase/firestore";
-import { app } from "@/services/firebase";
 import { QUERY_KEYS } from "@/constants/queryKeys";
+import { fetchRecentReviews } from "@/services/review/fetchRecentReviews";
 
-const db = getFirestore(app);
-
-export const useRecentReviews = () => {
-  return useQuery({
+export const useRecentReviews = () =>
+  useQuery({
     queryKey: [QUERY_KEYS.RECENT_REVIEWS],
-    queryFn: async () => {
-      const ref = collection(db, "reviews");
-      const q = query(ref, orderBy("createdAt", "desc"), limit(5));
-      const snapshot = await getDocs(q);
-      return snapshot.docs.map((doc) => ({
-        id: doc.id,
-        ...doc.data(),
-      }));
-    },
+    queryFn: fetchRecentReviews,
   });
-};
